requirements.txt
flask
flask-cors
python-dotenv
redis
pytest
PyJWT
